changing the err to double from int

+++ avc.cpp
@@ -85,7 +85,7 @@ int Robot::MeasureLine(){ //only coded for quad 2 rn
 			clock_gettime(CLOCK_MONOTONIC, &ts_end);
 			long dt = (ts_end.tv_sec-ts_start.tv_sec) * 1000000000 + ts_end.tv_nsec-ts_start.tv_nsec;
 			prev_error = line_error;
-			err = (int)((line_error*kp) + (((line_error - prev_error) * kd)/dt));
+			err = ((line_error*kp) + (((line_error - prev_error) * kd)/dt));
 					
 		    printf("\nwhiteness: %.1f",totwhite);
 		}
